fix: fix integer bug

+++ app/src/main/java/com/example/roommate/ui/fragments/FragmentCreateAd1.kt
@@ -39,7 +39,7 @@ class FragmentCreateAd1 : Fragment(R.layout.fragment_create_ad1) {
             title = binding.titleEt.text.toString(),
             rent_value =binding.rentEt.text.toString().takeIf { it.isNotEmpty() }?.toDouble() ?: 0.0,
             cond_value = binding.condEt.text.toString().takeIf { it.isNotEmpty() }?.toDouble() ?: 0.0,
-            max_client = binding.clientsEt.text.toString().takeIf { it.isNotEmpty() }?.toInt() ?: 0,
+            max_client = binding.clientsEt.text.toString().takeIf { it.isNotEmpty() }?.toLong() ?: 0,
             description = binding.descriptionEt.text.toString(),
             suite_qtt = null,
             bedroom_qtt = null,
